refactoring PR to use react router + left a comment to remove legacy code + refactoring component into separate components
